+++ packages/terra-core-docs/src/terra-dev-site/doc/divider/example/DividerWithText.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Divider from 'terra-divider';
+import SectionDivider from 'terra-divider/src/_SectionDivider';
 import DividerExampleTemplate from './DividerExampleTemplate';
 
 const currentDate = 'November 12, 1955';
 
 const DividerCustomExample = () => (
   <DividerExampleTemplate>
-    <Divider text={currentDate} />
+    <SectionDivider text={currentDate} level={3} />
   </DividerExampleTemplate>
 );