+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Added
+  * Added a Heading Level for `terra-Divider`.
+
+* Added
   * Added Accordion Examples for `terra-section-header`.
 
 ## 1.15.1 - (September 8, 2022)